Finish

+++ python_variable_annotations/6-sum_mixed_list.py
@@ -1,11 +1,10 @@
-from typing import List, Union
-
 #!/usr/bin/env python3
 '''
 strongly typed python
 '''
+from typing import List, Union
 
-def sum_mixed_list(input_list: List[Union[float, int]]) -> float:
+def sum_mixed_list(input_list: List[Union[int, float]]) -> float:
     '''
     function that adds 2 values
     '''
